requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
passlib[bcrypt]
python-multipart
python-jose
python-dotenv
aiosqlite
alembic